Merge from upstream

+++ salt/scripts.py
@@ -8,9 +8,13 @@ from __future__ import absolute_import, print_function
 import os
 import sys
 import time
+import signal
 import logging
+import functools
 import threading
 import traceback
+import signal
+import functools
 from random import randint
 
 # Import salt libs
@@ -38,6 +42,47 @@ def _handle_interrupt(exc, original_exc, hardfail=False, trace=''):
         raise exc
 
 
+def _handle_signals(client, signum, sigframe):
+    trace = traceback.format_exc()
+    try:
+        hardcrash = client.options.hard_crash
+    except (AttributeError, KeyError):
+        hardcrash = False
+
+    if signum == signal.SIGINT:
+        exit_msg = '\nExiting gracefully on Ctrl-c'
+        try:
+            jid = client.local_client.pub_data['jid']
+            exit_msg += (
+                '\n'
+                'This job\'s jid is: {0}\n'
+                'The minions may not have all finished running and any remaining '
+                'minions will return upon completion. To look up the return data '
+                'for this job later, run the following command:\n\n'
+                'salt-run jobs.lookup_jid {0}'.format(jid)
+            )
+        except (AttributeError, KeyError):
+            pass
+    else:
+        exit_msg = None
+
+    _handle_interrupt(
+        SystemExit(exit_msg),
+        Exception('\nExiting with hard crash on Ctrl-c'),
+        hardcrash, trace=trace)
+
+
+def _install_signal_handlers(client):
+    # Install the SIGINT/SIGTERM handlers if not done so far
+    if signal.getsignal(signal.SIGINT) is signal.SIG_DFL:
+        # No custom signal handling was added, install our own
+        signal.signal(signal.SIGINT, functools.partial(_handle_signals, client))
+
+    if signal.getsignal(signal.SIGTERM) is signal.SIG_DFL:
+        # No custom signal handling was added, install our own
+        signal.signal(signal.SIGINT, functools.partial(_handle_signals, client))
+
+
 def salt_master():
     '''
     Start the salt master.
@@ -51,8 +96,15 @@ def minion_process():
     '''
     Start a minion process
     '''
+    import salt.utils
     import salt.cli.daemons
+
     # salt_minion spawns this function in a new process
+
+    salt.utils.appendproctitle('KeepAlive')
+
+    def handle_hup(manager, sig, frame):
+        manager.minion.reload()
 
     def suicide_when_without_parent(parent_pid):
         '''
@@ -65,28 +117,35 @@ def minion_process():
             time.sleep(5)
             try:
                 # check pid alive (Unix only trick!)
-                os.kill(parent_pid, 0)
-            except OSError:
+                if os.getuid() == 0 and not salt.utils.is_windows():
+                    os.kill(parent_pid, 0)
+            except OSError as exc:
                 # forcibly exit, regular sys.exit raises an exception-- which
                 # isn't sufficient in a thread
+                log.error('Minion process encountered exception: {0}'.format(exc))
                 os._exit(salt.defaults.exitcodes.EX_GENERIC)
+
     if not salt.utils.is_windows():
         thread = threading.Thread(target=suicide_when_without_parent, args=(os.getppid(),))
         thread.start()
 
     minion = salt.cli.daemons.Minion()
+    signal.signal(signal.SIGHUP,
+                  functools.partial(handle_hup,
+                                    minion))
 
     try:
         minion.start()
     except (SaltClientError, SaltReqTimeoutError, SaltSystemExit) as exc:
-        log.warn('** Restarting minion **')
+        log.warning('Fatal functionality error caught by minion handler:\n', exc_info=True)
+        log.warning('** Restarting minion **')
         delay = 60
         if minion is not None and hasattr(minion, 'config'):
             delay = minion.config.get('random_reauth_delay', 60)
         delay = randint(1, delay)
         log.info('waiting random_reauth_delay {0}s'.format(delay))
         time.sleep(delay)
-        exit(salt.defaults.exitcodes.SALT_KEEPALIVE)
+        sys.exit(salt.defaults.exitcodes.SALT_KEEPALIVE)
 
 
 def salt_minion():
@@ -95,7 +154,10 @@ def salt_minion():
     Auto restart minion on error.
     '''
     import signal
-    import functools
+
+    import salt.utils.process
+    salt.utils.process.notify_systemd()
+
     import salt.cli.daemons
     import multiprocessing
     if '' in sys.path:
@@ -121,6 +183,8 @@ def salt_minion():
         os.kill(pid, signum)
 
     # keep one minion subprocess running
+    prev_sigint_handler = signal.getsignal(signal.SIGINT)
+    prev_sigterm_handler = signal.getsignal(signal.SIGTERM)
     while True:
         try:
             process = multiprocessing.Process(target=minion_process)
@@ -131,6 +195,9 @@ def salt_minion():
             signal.signal(signal.SIGINT,
                           functools.partial(escalate_signal_to_process,
                                             process.pid))
+            signal.signal(signal.SIGHUP,
+                          functools.partial(escalate_signal_to_process,
+                                            process.pid))
         except Exception:  # pylint: disable=broad-except
             # if multiprocessing does not work
             minion = salt.cli.daemons.Minion()
@@ -139,8 +206,13 @@ def salt_minion():
 
         process.join()
 
+        # Process exited or was terminated. Since we're going to try to restart
+        # it, we MUST, reset signal handling to the previous handlers
+        signal.signal(signal.SIGINT, prev_sigint_handler)
+        signal.signal(signal.SIGTERM, prev_sigterm_handler)
+
         if not process.exitcode == salt.defaults.exitcodes.SALT_KEEPALIVE:
-            break
+            sys.exit(process.exitcode)
         # ontop of the random_reauth_delay already preformed
         # delay extra to reduce flooding and free resources
         # NOTE: values are static but should be fine.
@@ -176,23 +248,28 @@ def proxy_minion_process(queue):
                 # forcibly exit, regular sys.exit raises an exception-- which
                 # isn't sufficient in a thread
                 os._exit(999)
+
     if not salt.utils.is_windows():
         thread = threading.Thread(target=suicide_when_without_parent, args=(os.getppid(),))
         thread.start()
 
     restart = False
     proxyminion = None
+    status = salt.defaults.exitcodes.EX_OK
     try:
         proxyminion = salt.cli.daemons.ProxyMinion()
         proxyminion.start()
     except (Exception, SaltClientError, SaltReqTimeoutError, SaltSystemExit) as exc:
         log.error('Proxy Minion failed to start: ', exc_info=True)
         restart = True
+        # status is superfluous since the process will be restarted
+        status = salt.defaults.exitcodes.SALT_KEEPALIVE
     except SystemExit as exc:
         restart = False
+        status = exc.code
 
     if restart is True:
-        log.warn('** Restarting proxy minion **')
+        log.warning('** Restarting proxy minion **')
         delay = 60
         if proxyminion is not None:
             if hasattr(proxyminion, 'config'):
@@ -203,6 +280,7 @@ def proxy_minion_process(queue):
         queue.put(random_delay)
     else:
         queue.put(0)
+    sys.exit(status)
 
 
 def salt_proxy_minion():
@@ -247,7 +325,7 @@ def salt_proxy_minion():
                 restart_delay = 60
             if restart_delay == 0:
                 # Minion process ended naturally, Ctrl+C, --version, etc.
-                break
+                sys.exit(process.exitcode)
             # delay restart to reduce flooding and allow network resources to close
             time.sleep(restart_delay)
         except KeyboardInterrupt:
@@ -264,6 +342,9 @@ def salt_syndic():
     '''
     Start the salt syndic.
     '''
+    import salt.utils.process
+    salt.utils.process.notify_systemd()
+
     import salt.cli.daemons
     pid = os.getpid()
     try:
@@ -278,20 +359,12 @@ def salt_key():
     Manage the authentication keys with salt-key.
     '''
     import salt.cli.key
-    client = None
     try:
         client = salt.cli.key.SaltKey()
+        _install_signal_handlers(client)
         client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    except Exception as err:
+        sys.stderr.write("Error: {0}\n".format(err))
 
 
 def salt_cp():
@@ -300,20 +373,9 @@ def salt_cp():
     master.
     '''
     import salt.cli.cp
-    client = None
-    try:
-        client = salt.cli.cp.SaltCPCli()
-        client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    client = salt.cli.cp.SaltCPCli()
+    _install_signal_handlers(client)
+    client.run()
 
 
 def salt_call():
@@ -324,20 +386,9 @@ def salt_call():
     import salt.cli.call
     if '' in sys.path:
         sys.path.remove('')
-    client = None
-    try:
-        client = salt.cli.call.SaltCall()
-        client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    client = salt.cli.call.SaltCall()
+    _install_signal_handlers(client)
+    client.run()
 
 
 def salt_run():
@@ -347,20 +398,9 @@ def salt_run():
     import salt.cli.run
     if '' in sys.path:
         sys.path.remove('')
-    client = None
-    try:
-        client = salt.cli.run.SaltRun()
-        client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    client = salt.cli.run.SaltRun()
+    _install_signal_handlers(client)
+    client.run()
 
 
 def salt_ssh():
@@ -370,20 +410,10 @@ def salt_ssh():
     import salt.cli.ssh
     if '' in sys.path:
         sys.path.remove('')
-    client = None
     try:
         client = salt.cli.ssh.SaltSSH()
+        _install_signal_handlers(client)
         client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
     except SaltClientError as err:
         trace = traceback.format_exc()
         try:
@@ -400,43 +430,37 @@ def salt_cloud():
     '''
     The main function for salt-cloud
     '''
+    # Define 'salt' global so we may use it after ImportError. Otherwise,
+    # UnboundLocalError will be raised.
+    global salt  # pylint: disable=W0602
+
     try:
+        # Late-imports for CLI performance
+        import salt.cloud
         import salt.cloud.cli
-        has_saltcloud = True
     except ImportError as e:
-        log.error("Error importing salt cloud {0}".format(e))
         # No salt cloud on Windows
-        has_saltcloud = False
+        log.error("Error importing salt cloud {0}".format(e))
+        print('salt-cloud is not available in this system')
+        sys.exit(salt.defaults.exitcodes.EX_UNAVAILABLE)
     if '' in sys.path:
         sys.path.remove('')
 
-    if not has_saltcloud:
-        print('salt-cloud is not available in this system')
-        sys.exit(salt.defaults.exitcodes.EX_UNAVAILABLE)
-
-    client = None
-    try:
-        client = salt.cloud.cli.SaltCloud()
-        client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    client = salt.cloud.cli.SaltCloud()
+    _install_signal_handlers(client)
+    client.run()
 
 
 def salt_api():
     '''
     The main function for salt-api
     '''
+    import salt.utils.process
+    salt.utils.process.notify_systemd()
+
     import salt.cli.api
     sapi = salt.cli.api.SaltAPI()  # pylint: disable=E1120
-    sapi.run()
+    sapi.start()
 
 
 def salt_main():
@@ -447,20 +471,9 @@ def salt_main():
     import salt.cli.salt
     if '' in sys.path:
         sys.path.remove('')
-    client = None
-    try:
-        client = salt.cli.salt.SaltCMD()
-        client.run()
-    except KeyboardInterrupt as err:
-        trace = traceback.format_exc()
-        try:
-            hardcrash = client.options.hard_crash
-        except (AttributeError, KeyError):
-            hardcrash = False
-        _handle_interrupt(
-            SystemExit('\nExiting gracefully on Ctrl-c'),
-            err,
-            hardcrash, trace=trace)
+    client = salt.cli.salt.SaltCMD()
+    _install_signal_handlers(client)
+    client.run()
 
 
 def salt_spm():
@@ -472,3 +485,17 @@ def salt_spm():
     import salt.cli.spm
     spm = salt.cli.spm.SPM()  # pylint: disable=E1120
     spm.run()
+
+
+def salt_extend(extension, name, description, salt_dir, merge):
+    '''
+    Quickstart for developing on the saltstack installation
+
+    .. versionadded:: 2016.11.0
+    '''
+    import salt.utils.extend
+    salt.utils.extend.run(extension=extension,
+                          name=name,
+                          description=description,
+                          salt_dir=salt_dir,
+                          merge=merge)
